web/pyproject.toml
[tool.poetry]
name = "mirra_backend"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
jinja2 = "^3.1.4"
uvicorn = "^0.30.6"
paho-mqtt = "^2.1.0"
pydantic-settings = "^2.4.0"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.34"
python-multipart = "^0.0.9"

[tool.poetry.group.dev]
optional=true

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
mypy = "^1.10.0"

[tool.poetry.group.test]
optional=true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "session"
addopts = "-rsxX -l --tb=short --strict-markers"
xfail_strict = true

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.mypy]
files = ["./mirra_backend/"]
mypy_path = "./mirra_backend/"
explicit_package_bases = false
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
